installs newer version of node
the yum repo only has an older version of node which, while it will run the app, won't run the integration tests. I changed the logic so it installs it from a downloaded binary. I also added psuedo-tests to check to make sure it installed correctly, and then restart jenkins after the install of gulp is complete.